fix: fix emqx cluster can not ready
